Removed Gaussian and identity. Greatly simplified.
Gaussian and identity kernels required a far more complex implementation, while random kernels can be implemented, roughly speaking, with a _requires_grad=False flag.
As gaussian and identity kernels performed quite poorly and are not the main focus of the project, we remove them completely from this branch and simplify the code greatly, making it easier to understand, support, extend, and reuse.